edited names, moved code to correct files as per project structure
