+++ routes/users.js
@@ -2,8 +2,9 @@ const express = require('express')
 
 const router = express.Router()
 const { validateUser } = require('../schemas/users')
-const { createUser } = require('../controllers/auth')
+const { postUser } = require('../controllers/users')
 
-router.post('/register', validateUser, createUser)
+// Falta agregar un middleware
+router.post('/register', validateUser, postUser)
 
 module.exports = router
